Implements Key-Value caching for Attention.
Designed to support fast autoregressive decoding by caching previously computed key and value tensors across multiple forward passes. Caching at the layer level is supported.

In addition to KV-Cache implementation, we add a simple testing harness as well.

A limitation of this change is that it Leaves types unchecked at runtime (in lieu, we describe shape annotations in docstrings).