Adds a Multiheaded Self Attention module with support for LoRA, RoPE, and KV Caching.
Includes a decomposition of `scaled_dot_product_attention` written by hand, for sake of learning. We emphasize that we preference fused kernel operations, which is a hint toward FlashAttention being dispatched under the hood. We also add some documentation on Non-rotary position embeddings, which allow the model to ignore position if it's not relevant.