Adds a Train Loop for a DeepSeek model + associated Integration Testing.
The file `train/shared.py` is intended to be consumed by a `pretrain.py` and a `sft.py` script in a followup commit. In addition to a training loop, we add associated utilities such as checkpointing, dataloading, and configuring a scheduler and optimizer. Although we do not unit test our training utilities, we do perform an integration test.