Adds support for Pretraining + SFT as training `mode`'s.
Note that we touched several `model/` files in order to ensure that our GPU devices had the requisite objects stored in memory for integration testing. Notably adds configuration files to describe (mock) model configuration as well as training arguments which can be useful for testing.

We do remark that we plan to "clean-up" the device-object management (i.e. placement of data on GPU) such that we can ensure we are not wasting resources and for ease of future maintenance.

For a link describing an example training run that was kicked off using _this_ configuration, please see: https://wandb.ai/asantucci-stanford-university/deepseek-pretrain/runs/yam47hey